+++ libraries/lib_userideas/Helper/PrepareItemAccessHelper.php
@@ -14,12 +14,12 @@ use Prism\Helper\HelperInterface;
 defined('JPATH_PLATFORM') or die;
 
 /**
- * This class provides functionality to prepare the access to the items.
+ * This class provides functionality to prepare an item access.
  *
  * @package      Userideas
  * @subpackage   Helpers
  */
-class PrepareAccess implements HelperInterface
+class PrepareItemAccessHelper implements HelperInterface
 {
     /**
      * User object.
@@ -39,40 +39,40 @@ class PrepareAccess implements HelperInterface
     }
 
     /**
-     * Prepare the access levels of the items.
+     * Prepare an item access level.
      *
-     * @param array $data
+     * @param \stdClass $data
      * @param array $options
      */
     public function handle(&$data, array $options = array())
     {
-        $userId = (int)$this->user->get('id');
-        $guest  = $this->user->get('guest');
-        $groups = $this->user->getAuthorisedViewLevels();
+        if (is_object($data)) {
+            $userId = (int)$this->user->get('id');
+            $guest  = $this->user->get('guest');
+            $groups = $this->user->getAuthorisedViewLevels();
 
-        foreach ($data as $key => $item) {
             // Compute the asset access permissions.
             // Technically guest could edit an article, but lets not check that to improve performance a little.
             if (!$guest) {
-                $asset = 'com_userideas.item.' . $item->id;
+                $asset = 'com_userideas.item.' . $data->id;
 
                 // Check general edit permission first.
                 if ($this->user->authorise('core.edit', $asset)) {
-                    $item->params->set('access-edit', true);
+                    $data->params->set('access-edit', true);
                 } // Now check if edit.own is available.
                 elseif ($userId > 0 and $this->user->authorise('core.edit.own', $asset)) {
                     // Check for a valid user and that they are the owner.
-                    if ($userId === (int)$item->user_id) {
-                        $item->params->set('access-edit', true);
+                    if ($userId === (int)$data->user_id) {
+                        $data->params->set('access-edit', true);
                     }
                 }
             }
 
             // If no access filter is set, the layout takes some responsibility for display of limited information.
-            if (!$item->catid or !$item->category_access) {
-                $item->params->set('access-view', in_array((int)$item->access, $groups, true));
+            if (!$data->catid or !$data->category_access) {
+                $data->params->set('access-view', in_array((int)$data->access, $groups, true));
             } else {
-                $item->params->set('access-view', in_array((int)$item->access, $groups, true) and in_array((int)$item->category_access, $groups, true));
+                $data->params->set('access-view', in_array((int)$data->access, $groups, true) and in_array((int)$data->category_access, $groups, true));
             }
         }
     }